requirements.txt
PySimpleGUI-4-foss==4.60.4.1
threading
cv2
ultralytics